+++ src/tables/property_table/property_table.js
@@ -17,13 +17,19 @@ angular.module('dashing.tables.property_table', [])
  *   style class to override
  * @param caption string (optional)
  *   the caption of the table
+ * @param caption-tooltip string (optional)
+ *   the tooltip of the caption
+ * @param sub-caption string (optional)
+ *   the table description
  *
  * @example
  *  <property-table
  *    props-bind="propsVariable"
  *    prop-name-class="col-md-4"
  *    prop-value-class="col-md-8"
- *    caption="Table caption">
+ *    caption="Table caption"
+ *    caption-tooltip="What is this table"
+ *    sub-caption="Table description">
  *  </property-table>
  */
   .directive('propertyTable', function() {
@@ -34,6 +40,8 @@ angular.module('dashing.tables.property_table', [])
       templateUrl: 'tables/property_table/property_table.html',
       scope: {
         caption: '@',
+        captionTooltip: '@',
+        subCaption: '@',
         props: '=propsBind',
         propNameClass: '@',
         propValueClass: '@'